+++ packages/cli/src/utils/clack_utils.ts
@@ -1,4 +1,5 @@
 import * as clack from '@clack/prompts'
+import * as fs from 'fs'
 import chalk from 'chalk'
 
 export async function abortIfCancelled<T>(input: T | Promise<T>): Promise<Exclude<T, symbol>> {
@@ -13,4 +14,14 @@ export async function abortIfCancelled<T>(input: T | Promise<T>): Promise<Exclud
 export async function abort(message?: string, status?: number): Promise<never> {
   clack.outro(message ?? chalk.red('setup cancelled.'))
   return process.exit(status ?? 1)
+}
+
+export function getNextRouter() {
+  if (fs.existsSync('src/app') || fs.existsSync('app')) {
+    return 'app'
+  } else if (fs.existsSync('src/pages') || fs.existsSync('pages')) {
+    return 'pages'
+  } else {
+    return undefined
+  }
 }
